refactor(gym): split tests_executor.rs (#1266)
tests_executor.rs 629 -> 319 LOC; new tests_executor_extra.rs.
